+++ src/PImageLoaderJpeg.hh
@@ -22,7 +22,7 @@
 namespace PImageLoaderJpeg
 {
     const char *getExt(void);
-    uchar* load(const std::string &file, uint &width, uint &height,
+    uchar* load(const std::string &file, size_t &width, size_t &height,
                 bool &use_alpha);
 }
